% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_players_names}
\alias{clean_players_names}
\title{Format player/item names as "Surname F."}
\usage{
clean_players_names(name)
}
\arguments{
\item{name}{character scalar or vector (e.g., "Roger Federer").}
}
\value{
Character vector with formatted names (e.g., "Federer R.").
}
\description{
Format player/item names as "Surname F."
}
\examples{
clean_players_names(c("Roger Federer", "Nadal"))
}
