% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{dirichlet_process_prior}
\alias{dirichlet_process_prior}
\title{Dirichlet-process prior for a latent partition}
\usage{
dirichlet_process_prior(concentration = 1)
}
\arguments{
\item{concentration}{Positive concentration parameter.}
}
\value{
An object of class \code{btsbm_partition_prior}.
}
\description{
Dirichlet-process prior for a latent partition
}
