% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.new_cluster_integral_log}
\alias{.new_cluster_integral_log}
\title{Gamma-Poisson integrated log-likelihood (new block)}
\usage{
.new_cluster_integral_log(wi, Zi, a_now, b_now)
}
\description{
Gamma-Poisson integrated log-likelihood (new block)
}
\keyword{internal}
