% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.new_cluster_integral_log_anchored}
\alias{.new_cluster_integral_log_anchored}
\title{Anchored version of the integrated log-score (drop a-only constant)}
\usage{
.new_cluster_integral_log_anchored(wi, Zi, a_now, b_now)
}
\description{
Anchored version of the integrated log-score (drop a-only constant)
}
\keyword{internal}
