% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expected_cl_py}
\alias{expected_cl_py}
\title{Expected number of clusters under finite/inf PY/DM-like priors (helper)}
\usage{
expected_cl_py(n, sigma, theta, K_DM)
}
\arguments{
\item{n}{integer(1) number of items.}

\item{sigma}{numeric(1) discount (0 for DP/DM).}

\item{theta}{numeric(1) concentration parameter.}

\item{K_DM}{integer(1) maximum number of clusters, or \code{Inf}.}
}
\value{
Numeric(1) expected number of clusters.
#' @references
Legramanti, S., Rigon, T., Durante, D., Dunson, D.B., 2022. Extended stochastic block models with application to criminal networks. The Annals of Applied Statistics 16. https://doi.org/10.1214/21-aoas1595
}
\description{
Expected number of clusters under finite/inf PY/DM-like priors (helper)
}
\keyword{internal}
