% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{gnedin_prior}
\alias{gnedin_prior}
\title{Gnedin prior for a latent partition}
\usage{
gnedin_prior(gnedin_hyperparameter = 0.5)
}
\arguments{
\item{gnedin_hyperparameter}{Gnedin hyperparameter in \verb{(0, 1)}.}
}
\value{
An object of class \code{btsbm_partition_prior}.
}
\description{
Gnedin prior for a latent partition
}
