% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lambda_to_theta}
\alias{lambda_to_theta}
\title{Map \eqn{\lambda} to Bradley–Terry \eqn{\theta = \lambda_i / (\lambda_i + \lambda_j)}}
\usage{
lambda_to_theta(lambda)
}
\arguments{
\item{lambda}{Numeric vector of positive rates \eqn{\lambda_i}.}
}
\value{
A matrix with entries \eqn{\theta_{ij} = \lambda_i / (\lambda_i + \lambda_j)}.
Diagonal is 1/2 by the formula; you may overwrite if you prefer NA on the diagonal.
}
\description{
Map \eqn{\lambda} to Bradley–Terry \eqn{\theta = \lambda_i / (\lambda_i + \lambda_j)}
}
\examples{
lambda_to_theta(c(1,2,4))
}
