% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{loo_btsbm}
\alias{loo_btsbm}
\title{Run Pareto-smoothed importance-sampling LOO for a fit}
\usage{
loo_btsbm(fit, ...)
}
\arguments{
\item{fit}{A \code{btsbm_fit} object.}

\item{...}{Passed to \code{\link[loo:loo]{loo::loo()}}.}
}
\value{
A \code{loo} object.
}
\description{
Run Pareto-smoothed importance-sampling LOO for a fit
}
