% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{partition_moves}
\alias{partition_moves}
\title{Set partition-move controls}
\usage{
partition_moves(
  split_merge = 0L,
  ranking_split_merge = 0L,
  item_split_merge = 0L,
  pair_proposal = c("uniform", "balanced", "informed"),
  restricted_scans = 1L
)
}
\arguments{
\item{split_merge}{Number of generic split--merge moves per iteration.}

\item{ranking_split_merge, item_split_merge}{PL--LBM moves for the ranking
and item partitions respectively.}

\item{pair_proposal}{Pair-selection rule for a future split--merge kernel.}

\item{restricted_scans}{Number of restricted Gibbs scans.}
}
\value{
An object of class \code{btsbm_partition_moves}.
}
\description{
Set partition-move controls
}
