% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{pitman_yor_prior}
\alias{pitman_yor_prior}
\title{Pitman--Yor prior for a latent partition}
\usage{
pitman_yor_prior(concentration = 1, discount = 0.2)
}
\arguments{
\item{concentration}{Concentration parameter.}

\item{discount}{Discount parameter in \verb{[0, 1)}.}
}
\value{
An object of class \code{btsbm_partition_prior}.
}
\description{
Pitman--Yor prior for a latent partition
}
