% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{pl_identifiability}
\alias{pl_identifiability}
\title{Set the identifiable representation for Plackett--Luce abilities}
\usage{
pl_identifiability(method = c("logmean0", "none"))
}
\arguments{
\item{method}{Normalisation method.  \code{"logmean0"} gives each reported
ability vector geometric mean one.}
}
\value{
An object of class \code{btsbm_pl_identifiability}.
}
\description{
Set the identifiable representation for Plackett--Luce abilities
}
