% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_models.R
\name{pl_model}
\alias{pl_model}
\title{Define a simple Plackett--Luce model}
\usage{
pl_model(
  latent_strength = .default_latent_strength(),
  identifiability = pl_identifiability(),
  ability_prior = NULL
)
}
\arguments{
\item{latent_strength}{Gamma prior for positive item strengths.}

\item{identifiability}{PL reporting constraint.}

\item{ability_prior}{Deprecated alias for \code{latent_strength}.}
}
\value{
A \code{btsbm_model}.
}
\description{
Define a simple Plackett--Luce model
}
