% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{posterior_similarity}
\alias{posterior_similarity}
\title{Compute posterior similarity for an inferred partition}
\usage{
posterior_similarity(fit, target = c("item", "ranking"))
}
\arguments{
\item{fit}{A \code{btsbm_fit} object.}

\item{target}{Either \code{"item"} or \code{"ranking"}.}
}
\value{
A posterior similarity matrix.
}
\description{
Compute posterior similarity for an inferred partition
}
