% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{posterior_strength}
\alias{posterior_strength}
\title{Summarise posterior item strengths}
\usage{
posterior_strength(fit, summary = c("draws", "mean"))
}
\arguments{
\item{fit}{A simple BT/PL or item-SBM fit.}

\item{summary}{Either \code{"draws"} or \code{"mean"}.}
}
\value{
A draws matrix or named posterior mean vector.
}
\description{
Summarise posterior item strengths
}
