% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.btsbm_fit}
\alias{print.btsbm_fit}
\title{Print a BTSBM fit}
\usage{
\method{print}{btsbm_fit}(x, ...)
}
\arguments{
\item{x}{A \code{btsbm_fit} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a BTSBM fit
}
