% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon entropy (nats)}
\usage{
shannon_entropy(p)
}
\arguments{
\item{p}{numeric vector of nonnegative masses.}
}
\value{
numeric(1) entropy in nats.
}
\description{
Shannon entropy (nats)
}
\keyword{internal}
