% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{summary.btsbm_fit}
\alias{summary.btsbm_fit}
\title{Summarise a BTSBM fit}
\usage{
\method{summary}{btsbm_fit}(object, ...)
}
\arguments{
\item{object}{A \code{btsbm_fit} object.}

\item{...}{Unused.}
}
\value{
A list containing model and sampler summaries.
}
\description{
Summarise a BTSBM fit
}
