% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urn_DP}
\alias{urn_DP}
\title{Urn weight: Dirichlet Process}
\usage{
urn_DP(v_minus, alpha_PY)
}
\arguments{
\item{v_minus}{integer sizes of occupied clusters (excluding focal item).}

\item{alpha_PY}{numeric concentration (>0).}
}
\value{
Numeric vector length \eqn{K+1}: existing weights and new-cluster mass.
#' @references
Legramanti, S., Rigon, T., Durante, D., Dunson, D.B., 2022. Extended stochastic block models with application to criminal networks. The Annals of Applied Statistics 16. https://doi.org/10.1214/21-aoas1595
}
\description{
Urn weight: Dirichlet Process
}
\keyword{internal}
