% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urn_PY}
\alias{urn_PY}
\title{Urn weight: Pitman–Yor process}
\usage{
urn_PY(v_minus, alpha_PY, sigma_PY)
}
\arguments{
\item{v_minus}{integer sizes of occupied clusters (excluding focal item).}

\item{alpha_PY}{numeric concentration (> -sigma_PY).}

\item{sigma_PY}{numeric discount in [0,1).}
}
\value{
Numeric vector length \eqn{K+1}.
}
\description{
Urn weight: Pitman–Yor process
}
\keyword{internal}
