% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_helper.R, R/utils.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Shannon entropy of a discrete distribution}
\usage{
shannon_entropy(p)

shannon_entropy(p)
}
\arguments{
\item{p}{numeric vector of nonnegative masses.}
}
\value{
numeric(1) entropy in nats.

numeric(1) entropy in nats.
}
\description{
Shannon entropy of a discrete distribution

Shannon entropy (nats)
}
\keyword{internal}
