% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urn_DM}
\alias{urn_DM}
\title{Urn weight: Dirichlet–Multinomial with max H_DM clusters}
\usage{
urn_DM(v_minus, beta_DM, H_DM)
}
\arguments{
\item{v_minus}{integer sizes of occupied clusters (excluding current item).}

\item{beta_DM}{numeric concentration.}

\item{H_DM}{integer maximum number of clusters.}
}
\value{
Numeric vector of length \eqn{H+1}: existing weights and new-cluster mass.
}
\description{
Urn weight: Dirichlet–Multinomial with max H_DM clusters
}
\keyword{internal}
