% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urn_DP}
\alias{urn_DP}
\title{Urn weight: Dirichlet Process}
\usage{
urn_DP(v_minus, alpha_PY)
}
\arguments{
\item{v_minus}{integer sizes of occupied clusters (excluding current item).}

\item{alpha_PY}{numeric concentration.}
}
\value{
Numeric vector of length \eqn{H+1}: existing weights and new-cluster mass.
}
\description{
Urn weight: Dirichlet Process
}
\keyword{internal}
