% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urn_GN}
\alias{urn_GN}
\title{Urn weight: Gnedin heuristic}
\usage{
urn_GN(v_minus, gamma)
}
\arguments{
\item{v_minus}{integer sizes of occupied clusters (excluding focal item).}

\item{gamma}{numeric parameter (>0).}
}
\value{
Numeric vector length \eqn{K+1}.
}
\description{
Urn weight: Gnedin heuristic
}
\keyword{internal}
