% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{urn_PY}
\alias{urn_PY}
\title{Urn weight: Pitman–Yor process}
\usage{
urn_PY(v_minus, alpha_PY, sigma_PY)
}
\arguments{
\item{v_minus}{integer sizes of occupied clusters (excluding focal item).}

\item{alpha_PY}{numeric concentration (> -sigma_PY).}

\item{sigma_PY}{numeric discount in [0,1).}
}
\value{
Numeric vector length \eqn{K+1}.
#' @references
Legramanti, S., Rigon, T., Durante, D., Dunson, D.B., 2022. Extended stochastic block models with application to criminal networks. The Annals of Applied Statistics 16. https://doi.org/10.1214/21-aoas1595
}
\description{
Urn weight: Pitman–Yor process
}
\keyword{internal}
